FIX: Critical dtype bug in Phase 1 dataset generation + investigation
Root Cause:
- geometric_metrics.py:285 created low_entropy pattern without dtype=np.uint8
- NumPy defaulted to int32, causing 7× slowdown vs uint8
- Created false impression that repetitive patterns are pathological

Impact on Phase 1 Findings:
- INCORRECT: "Repetitive patterns 40× slower" (was dtype bug)
- CORRECT: Repetitive patterns only 1.3× slower (normal variance)
- Simplifies Phase 2: no special handling for repetitive patterns needed

Bug Fix:
- Added dtype=np.uint8 to low_entropy pattern generation
- Regenerated low_entropy_1M.npy with correct dtype
- All datasets now consistent (uint8)

Investigation:
- Created 3 systematic investigation scripts
- Isolated dtype as root cause (7× impact)
- Confirmed pattern type has minimal impact (1.3×)
- Documented findings in REPETITIVE_PATTERN_INVESTIGATION.md

Files Changed:
- benchmarks/utils/geometric_metrics.py (bug fix: line 285)
- benchmarks/datasets/synthetic/low_entropy_1M.npy (regenerated)
- benchmarks/investigate_repetitive_performance.py (investigation)
- benchmarks/analyze_phase1_datasets.py (dataset analysis)
- benchmarks/test_0121_pattern.py (dtype isolation)
- reports/REPETITIVE_PATTERN_INVESTIGATION.md (comprehensive report)

Performance Corrected:
- Low entropy (int32 BUG): 455 Mops/s → (uint8 FIX): 2,776 Mops/s
- Ratio vs random: 40× → 1.28× (within variance)

Lessons Learned:
- ALWAYS verify dtype=np.uint8 in benchmark datasets
- Reproduce findings before drawing major conclusions
- Isolate variables systematically (dtype, pattern, size)

+++ benchmarks/utils/geometric_metrics.py
@@ -282,7 +282,7 @@ def generate_synthetic_dataset(entropy_level: str, size: int = 1000000) -> Tuple
 
     if entropy_level == 'low':
         # Low entropy: Highly repetitive pattern
-        pattern = np.array([0, 1, 2, 1] * 100)  # 400-element pattern
+        pattern = np.array([0, 1, 2, 1] * 100, dtype=np.uint8)  # 400-element pattern
         repetitions = size // len(pattern) + 1
         data = np.tile(pattern, repetitions)[:size]
